tidy of docker

+++ scripts/docker/nginx-oss/entrypoint.sh
@@ -21,8 +21,10 @@ echo "starting nginx ..."
 
 nginx_pid=$!
 
-cp /agent/nginx-agent.conf /etc/nginx-agent/nginx-agent.conf
-cat /etc/nginx-agent/nginx-agent.conf
+[ -e "/agent/nginx-agent.conf" ] && cp /agent/nginx-agent.conf /etc/nginx-agent/nginx-agent.conf
+ cat /etc/nginx-agent/nginx-agent.conf;
+
+[ -e "/agent/nginx.conf" ] && cp /agent/nginx.conf /etc/nginx/nginx.conf
 
 # start nginx-agent, pass args
 echo "starting nginx-agent ..."
